refactor(command-visualizer): split jumbotron.ts under 250-line limit
Split the 496-line jumbotron.ts into cohesive render/ siblings:
- jumbotron.ts keeps the markup half (commandBar/metadataBlock/dots/
  serializeJumbotron) and re-exports jumbotronCss for a stable public surface.
- jumbotron-css.ts: jumbotronCss assembler + baseCss/themedCss/carouselTrack
  section builders.
- jumbotron-css-keyframes.ts: the dot + command-pill @keyframes builder.
- jumbotron-shared.ts: NL, frameCommands, timelinePct, commandFrameIndices —
  shared by both halves (no cycle, no duplication; fallow clean).
Rendered CSS + markup byte-identical.
